backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
Pillow>=10.0.0
django-celery-results>=2.5.0
pytest>=7.0.0
pytest-django>=4.0.0
pytest-cov>=4.1.0
